Migrator for ambra v0.9 Citations. See the applicationContext.xml for tuning information. By default this runs in a back-ground thread allowing web-traffic to go thru. The migrations are cache aware and so there is no need to restart peers after the migrations are done.
If 'background' strategy is chosen, tune the txn time outs
and blobThrottle to adjust maximum concurrency. The migration
requires a write txn from mulgara - and so it is not expected
that ingests etc. are attempted till migrations are all complete.
Also note that a mulgara txn related error will terminate the
migration operation and a restart of ambra is required to continue
the rest of the migrations.

If 'background' strategy is not chosen, ambra start up will succeed
only when all Citations in the database are  migrated.

Addresses #921


git-svn-id: http://svn.topazproject.org/svn/head@6953 48088309-b40f-0410-b149-b14953253675