Search Indexer for all @Searchable annotated properties. Please see applicationContext.xml for tuning.
By default it runs in the background allowing web-traffic
to go thru. Since this requires a write txn from mulgara,
adjust the throttles and txn timeout to allow maximum
concurrency.

If 'background' option is not chosen, ambra startup succeeds
only when all indexable fields are indexed.

Also by default, 'finalize' is false. It is recommended to
set this to true in production to avoid un-necessary
re-indexing of all new data at start-up.


git-svn-id: http://svn.topazproject.org/svn/head@6954 48088309-b40f-0410-b149-b14953253675