This commit is related to #520
DNA sequences listed in a paper are often lengthy, as in this example from PLoS ONE:

http://www.plosone.org/doi/pone.0000299#s4

The string of letters is not wrapping at the end of the line, it just extends off to edge of the screen, behind other layers. 

The XSL sheet will now search for words over 40 characters and insert a zero width space to allow the browser to wrap text when needed.



git-svn-id: http://svn.topazproject.org/svn/head@7011 48088309-b40f-0410-b149-b14953253675