This commit is related to #1068
When viewing figures & tables in the 'slideshow' within articles, some terms within the title or caption that should be italicized are not. it looks like <italic> is preserved, but <named-content> is not. 

This change fixes this.  The change in the XSL forces the output to be formated correctly, the change in the CSS tells the browser how to handle style the text.

git-svn-id: http://svn.topazproject.org/svn/head@7013 48088309-b40f-0410-b149-b14953253675