This is a recommit for Ticket #1071 with new updated formatting.  The problem was an incorrect number of tabs on lines 3725, 3726, and 3728.  These tabs were not affecting the build in anyway as I would have caught them in my tests before checkin.
Now for the actual fix, I needed to make some modifications to how the 'references' link is pulled into the side nav bar toc on article pages.

Specifically:

(a) make sure that a 'references' link is only pulled into the side nav article toc for <ref-list> that's within the <back> section. we have cases of <ref-list> within the article body (eg, an 'additional reading' list within boxed-text), and we don't want those to appear on the side nav toc.

(b) make sure that for 'references' links that do get pulled in, we pull the contents of the ref-list <title>, rather than just always using the text 'references'. we have cases where the ref-list is entitled 'further reading' or somesuch; as these aren't technically reference lists, and as we want the side nav toc link to be consistent with the actual heading, we need to pull in the <title> if there is one, and default to 'references' if there is no <title>.

git-svn-id: http://svn.topazproject.org/svn/head@7018 48088309-b40f-0410-b149-b14953253675