Fixed the replay transaction log problem using run-tql tool.
The problem was because of a invalid code associated with trying to concatenate
multiple lines till a ';' is encountered to mark end of query.  The code made
the assumption that the only ';' encountered will be marking the end of the
line and not in the data itself. The specific case here had ';' in the value of
the eLocationId causing the code to try and execute an invalid query.
Unfortunately, the break was right around UTF-8 encoded character and thus the
confusion on why the lexer was having problems.

The fix here is not the best, but avoids having to implement a complete
understanding of the TQL syntax in this tool.

WARNING: This fix breaks the assumption that you can have two queries in the
same line. I don't believe we are using this anywhere, but thought a warning
should be provided anyway.

git-svn-id: http://svn.topazproject.org/svn/head@7144 48088309-b40f-0410-b149-b14953253675