Re-work the query for listing flagged annotations/replies on admin page to improve performance.
The annotea:body used previously as the discriminator for Annotea instances
had the unintended side effect of getting the journal filters applied.
While this in itself is not a problem, currently the application of filter
queries for Oql is not subject to the optimization done in r7156 and
therefore requires improved handling of <mulgara:is> in mulgara.

btw: The reason for adding further constraints on 'Annotea' in the first place
is because Annotea does not define any rdf:type values.


git-svn-id: http://svn.topazproject.org/svn/head@7161 48088309-b40f-0410-b149-b14953253675