adding ambraJournal to journals
in 0.9, we moved config.ambra.freemarker out of ambra-webapp's defaults.xml and into the overlay journal's defaults.xml

the idea was that the defaultJournal would contain the default freemarker config for the whole system.

at that point, we should have created an ambraJournal under ambra/journals so that things would work properly in a non-plos build, but since almost all the devs were working in a plos context, we skipped it.

in 0.9.1, we've made some changes that require the default journal to be distinct from any of the "real" overlay journals in order for some override functionality to work. we really need to have a separate default journal under ambra/journals now, so it there's some additional reason to create ambra/journals/journal-ambraJournal

in addition to adding journal-ambraJournal, i moved many default settings from journal-plosone into journal-ambraJournal, and updated the plos-specific packages/ambra/../ambra.xml to use AmbraJournal as its default instead of PLoSJournal (hopefully we'll get rid of the plos-specific config here in 0.9.2)

git-svn-id: http://svn.topazproject.org/svn/head@7248 48088309-b40f-0410-b149-b14953253675