reverse order of mov and qt extensions, as mov is more generic that qt.
when an asset is placed in fedora, the filename extension is stripped and the mime type is stored. mime.types handles this translation.

then, when the asset is served ambra, a filename extension is added based on the mime type.

in the case where multiple extensions maps to a single mime type, ambra chooses the first extension on the list. in this case, mov is a better default choice than qt.

git-svn-id: http://svn.topazproject.org/svn/head@7286 48088309-b40f-0410-b149-b14953253675