Implemented dynamic display of recent articles on the front page of every Journal.
Each time this page is displayed, the "Recent Articles" section is populated (in random order) by random selection from a pool of the most recent articles published in that Journal.  Display is controled by the Configuration properties "numDaysInPast", "numArticlesToShow", and "typeUriArticlesToShow", all of which are set in the defaults.xml file for each Journal.  Removed static file recentArticles.ftl from each of the Journals.  Edited file index.ftl (for each Journal) to point to the default (dynamic) recentArticles.ftl file.

Addresses #1052
Addresses #1076


git-svn-id: http://svn.topazproject.org/svn/head@7306 48088309-b40f-0410-b149-b14953253675