avoid calling articleInfoX variable without first checking that it exists.
create new variable, docTitle, in article_variables.ftl, and remove all references to articleInfoX in various templates.

often, when caches are out of whack or other errors occur, articleInfoX is empty. however, we can still display most of the article using the xsl transform, and it's nicer to display what we can than to fail completely with a freemarker error.

git-svn-id: http://svn.topazproject.org/svn/head@7379 48088309-b40f-0410-b149-b14953253675