Make changes to admin panel to allow the ordering of articles within issues. The admin panel interface was updated to allow the ordering of articles and increase the usability and reduce errors in certain task.
Additional Management Panels/Actions

 1. Cross Publishing
 2. Volume Management
 3. Issue Management

Panels Changed
 1. Manage Virtual Journals

Panels Deleted
 1. Manage Volumes/Issues and its Templete ftl

New AdminService provides functionality that can be used by all the admin panels.

The Issue.java model has been changed so that simpleCollections are not used and ordering of articles is maintained in a separate list. This will be refactored out into a separate PloS issue and volume classes in the next development cycle so that PloS specific concerns are removed.

Addresses #1109

git-svn-id: http://svn.topazproject.org/svn/head@7422 48088309-b40f-0410-b149-b14953253675