Prevent Cross Publishing into current journal.
After and ingest the admin is presented with a matrix of journals that articles can be cross published in. Included in the current journal the article will be published in if 'Publish' check box is chosen. Articles should not be crossed published in the journal they are published in. Remove it form the adminTop.ftl but several xxxxAction.java files were affect so they were changed also.

Addresses #1149

git-svn-id: http://svn.topazproject.org/svn/head@7438 48088309-b40f-0410-b149-b14953253675