Display authors in search results in the order they appear in the article.
DublinCore.creators is currently unordered; while it would be possible to
change it to be ordered, that would not come for free in terms of query
performance. Using the already ordered author list from the citation is a
bit slower, but not hugely so: it slows down retrieval of the article
info for a search hit by 10% (i.e. a page full of results now takes 10s
instead of 9s).

This closes #1124.

git-svn-id: http://svn.topazproject.org/svn/head@7456 48088309-b40f-0410-b149-b14953253675