Article page, right-hand column: now use a non-URL-encoded URL when submitting the Article to the Digg social bookmarking site.  This is an intermittent bug that may show up again, since it seems to be an issue with Digg.
The core problem was that the URL submitted to Digg was being URL-encoded *twice*.  However, it seems that double-URL-encoding was being performed by the Digg javascript and so was not dependent on Ambra code.  Ambra now does not URL-encode the URL at all, just in case that might help, but extensive testing indicated that even this measure is not enough to prevent occassional double-URL-encoding of submitted URLs.

Since this issue seems to reside with Digg itself, it is unlikely to be corrected by any further changes to the Ambra code.

Addresses #1181


git-svn-id: http://svn.topazproject.org/svn/head@7519 48088309-b40f-0410-b149-b14953253675