Fix overly aggressive optimization in r7266.
The problem is if a field has multiple values (i.e. cardinality greater 1)
then "search(a.b, '+x +y')" is not the same as
"search(a.b, 'x') and search(a.b, 'y')" if 'x' and 'y' appear in different
values of field 'b'. I.e. the collapsing of terms into a single search is
only valid for optional terms (factors of a disjunction), not for required
terms (factors of a conjuction).

This fixes #1176.

git-svn-id: http://svn.topazproject.org/svn/head@7537 48088309-b40f-0410-b149-b14953253675